@@ -86,28 +89,32 @@
     "negotiator": "^0.6.2",
     "nunjucks": "^3.2.3",
     "otplib": "^12.0.1",
+    "pcf-start": "^1.18.4",
     "redis": "^3.1.2",
     "require-directory": "^2.1.1",
     "serve-favicon": "^2.5.0",
     "session-file-store": "^1.5.0",
     "supertest": "^6.2.2",
     "toobusy-js": "^0.5.1",
+    "to-boolean": "^1.0.0",
     "ts-node": "^10.4.0",
-    "tsconfig-paths": "^3.11.0",
-    "typescript": "^4.5.4",
+    "tsconfig-paths": "^4.1.0",
+    "typescript": "^4.7.4",
     "uuid": "^8.3.2",
-    "yarn": "^1.22.18"
+    "yarn": "^1.22.18",
+    "launchdarkly-node-server-sdk": "^6.3.0",
+    "ldclient-node": "^5.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.10",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-typescript": "^7.16.7",
     "@types/copy-webpack-plugin": "^8.0.1",
-    "@types/jest": "^27.4.1",
+    "@types/jest": "^27.5.0",
     "@types/mini-css-extract-plugin": "^2.5.1",
     "@types/webpack-dev-middleware": "^5.0.2",
     "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@typescript-eslint/parser": "^5.31.0",
     "@wdio/sauce-service": "^7.16.13",
     "allure-commandline": "^2.13.8",
     "axios-debug-log": "^0.8.4",
