@@ -1,10 +1,9 @@
 module.exports = {
   roots: ['<rootDir>/src/test/a11y'],
-  testRegex: '(/src/test/.*|\\.(test|spec))\\.(ts|js)$',
+  testRegex: '(/src/test/.*|\\.test)\\.(ts|js)$',
   moduleFileExtensions: ['ts', 'js'],
   testEnvironment: 'node',
   transform: {
-    '^.+\\.ts?$': 'ts-jest',
+    '^.+\\.ts$': 'ts-jest',
   },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
 };
