@@ -3,7 +3,9 @@ import config from 'config';
 import jwt_decode from 'jwt-decode';
 
 import { PageLink } from '../../../steps/urls';
-import { UserDetails } from '../../controller/AppRequest';
+import { CosApiClient } from '../../case/CosApiClient';
+import { CaseWithId } from '../../case/case';
+import { AppRequest, UserDetails } from '../../controller/AppRequest';
 
 export const getRedirectUrl = (serviceUrl: string, callbackUrlPageLink: PageLink): string => {
   const id: string = config.get('services.idam.clientID');
@@ -19,15 +21,19 @@ export const getUserDetails = async (
   callbackUrlPageLink: PageLink
 ): Promise<UserDetails> => {
   const id: string = config.get('services.idam.clientID');
-  const secret: string = config.get('services.idam.clientSecret');
+  const secret: string = config.get('services.idam.citizenClientSecret');
   const tokenUrl: string = config.get('services.idam.tokenURL');
   const callbackUrl = encodeURI(serviceUrl + callbackUrlPageLink);
+  console.log('******* Trying to get user details');
+
   const code = encodeURIComponent(rawCode);
   const data = `client_id=${id}&client_secret=${secret}&grant_type=authorization_code&redirect_uri=${callbackUrl}&code=${code}`;
   const headers = { Accept: 'application/json', 'Content-Type': 'application/x-www-form-urlencoded' };
   const response: AxiosResponse<OidcResponse> = await Axios.post(tokenUrl, data, { headers });
-  const jwt = jwt_decode(response.data.id_token) as IdTokenJwtPayload;
-
+  console.log('******* Recieved response');
+  const jwt: IdTokenJwtPayload = jwt_decode(response.data.id_token);
+  console.log('******* Recieved jwt');
+  console.log('******* Recieved user: ');
   return {
     accessToken: response.data.access_token,
     id: jwt.uid,
@@ -39,19 +45,15 @@ export const getUserDetails = async (
 
 export const getSystemUser = async (): Promise<UserDetails> => {
   const id: string = config.get('services.idam.clientID');
-  const secret: string = config.get('services.idam.clientSecret');
+  const secret: string = config.get('services.idam.citizenClientSecret');
   const tokenUrl: string = config.get('services.idam.tokenURL');
-
   const systemUsername: string = config.get('services.idam.systemUsername');
   const systemPassword: string = config.get('services.idam.systemPassword');
 
   const headers = { Accept: 'application/json', 'Content-Type': 'application/x-www-form-urlencoded' };
-  const data = `grant_type=password&username=${systemUsername}&password=${systemPassword}&client_id=${id}
-                &client_secret=${secret}&scope=openid%20profile%20roles%20openid%20roles%20profile`;
-
+  const data = `grant_type=password&username=${systemUsername}&password=${systemPassword}&client_id=${id}&client_secret=${secret}&scope=openid%20profile%20roles%20openid%20roles%20profile`;
   const response: AxiosResponse<OidcResponse> = await Axios.post(tokenUrl, data, { headers });
-  const jwt = jwt_decode(response.data.id_token) as IdTokenJwtPayload;
-
+  const jwt: IdTokenJwtPayload = jwt_decode(response.data.id_token);
   return {
     accessToken: response.data.access_token,
     id: jwt.uid,
@@ -59,6 +61,14 @@ export const getSystemUser = async (): Promise<UserDetails> => {
     givenName: jwt.given_name,
     familyName: jwt.family_name,
   };
+};
+
+export const getCaseDetails = async (req: AppRequest): Promise<CaseWithId[]> => {
+  console.log('******* retrieveing the case details');
+  const cosApiClient = new CosApiClient(req.session.user.accessToken, 'http://localhost:3001');
+  const response = await cosApiClient.retrieveCasesByUserId(req.session.user);
+  console.log('******* got the response for case details');
+  return response;
 };
 
 interface IdTokenJwtPayload {
