@@ -1,10 +1,15 @@
+// eslint-disable-next-line import/no-unresolved
 import autobind from 'autobind-decorator';
 import { Response } from 'express';
 
 import { getNextStepUrl } from '../../steps';
+import { ApplicantUploadFiles, RespondentUploadFiles, UploadDocumentSucess } from '../../steps/constants';
 import { RESPONDENT_TASK_LIST_URL, SAVE_AND_SIGN_OUT } from '../../steps/urls';
+import { getSystemUser } from '../auth/user/oidc';
+import { getCaseApi } from '../case/CaseApi';
+import { CosApiClient } from '../case/CosApiClient';
 import { Case, CaseWithId } from '../case/case';
-import { CITIZEN_SAVE_AND_CLOSE, CITIZEN_UPDATE, State } from '../case/definition';
+import { CITIZEN_SAVE_AND_CLOSE, CITIZEN_UPDATE, CaseData, State } from '../case/definition';
 import { Form, FormFields, FormFieldsFn } from '../form/Form';
 import { ValidationError } from '../form/validation';
 
@@ -13,9 +18,7 @@ import { AppRequest } from './AppRequest';
 @autobind
 export class PostController<T extends AnyObject> {
   //protected ALLOWED_RETURN_URLS: string[] = [CHECK_ANSWERS_URL];
-
   constructor(protected readonly fields: FormFields | FormFieldsFn) {}
-
   /**
    * Parse the form body and decide whether this is a save and sign out, save and continue or session time out
    */
@@ -31,6 +34,9 @@ export class PostController<T extends AnyObject> {
       await this.saveBeforeSessionTimeout(req, res, formData);
     } else if (req.body.accessCodeCheck) {
       await this.checkCaseAccessCode(req, res, form, formData);
+      await this.getCaseList(req, res, form, formData);
+    } else if (req.body.onlyContinue) {
+      await this.onlyContinue(req, res, form, formData);
     } else {
       await this.saveAndContinue(req, res, form, formData);
     }
@@ -57,16 +63,22 @@ export class PostController<T extends AnyObject> {
   private async saveAndContinue(req: AppRequest<T>, res: Response, form: Form, formData: Partial<Case>): Promise<void> {
     Object.assign(req.session.userCase, formData);
     req.session.errors = form.getErrors(formData);
-
+    console.log('errors are:', req.session.errors);
     this.filterErrorsForSaveAsDraft(req);
 
     if (req.session.errors.length) {
       return this.redirect(req, res);
     }
 
-    req.session.userCase = await this.save(req, formData, this.getEventName(req));
+    if (req.originalUrl.includes(UploadDocumentSucess)) {
+      if (req?.session?.userCase?.applicantUploadFiles) {
+        req.session.userCase[ApplicantUploadFiles] = [];
+      }
+      if (req?.session?.userCase?.respondentUploadFiles) {
+        req.session.userCase[RespondentUploadFiles] = [];
+      }
+    }
 
-    //this.checkReturnUrlAndRedirect(req, res, this.ALLOWED_RETURN_URLS);
     this.redirect(req, res);
   }
 
@@ -84,9 +96,37 @@ export class PostController<T extends AnyObject> {
 
   protected async save(req: AppRequest<T>, formData: Partial<Case>, eventName: string): Promise<CaseWithId> {
     try {
-      console.log(eventName);
       Object.assign(req.session.userCase, formData);
-      // req.session.userCase = await req.locals.api.triggerEvent(req.session.userCase.id, formData, eventName);
+      // call here to get the case details //
+      const citizenUser = req.session.user;
+      req.locals.api = getCaseApi(citizenUser, req.locals.logger);
+      const caseReference = req.session.userCase.caseCode;
+      const caseData = await req.locals.api.getCaseById(caseReference as string);
+      console.log('Saving data for case : ' + JSON.stringify(caseData.id));
+      req.session.userCase = await req.locals.api.triggerEvent(req.session.userCase.id, formData, eventName);
+    } catch (err) {
+      req.locals.logger.error('Error saving', err);
+      req.session.errors = req.session.errors || [];
+      req.session.errors.push({ errorType: 'errorSaving', propertyName: '*' });
+    }
+    return req.session.userCase;
+  }
+
+  protected async saveData(
+    req: AppRequest<T>,
+    formData: Partial<Case>,
+    eventName: string,
+    data: Partial<CaseData>
+  ): Promise<CaseWithId> {
+    try {
+      console.log(eventName);
+
+      req.session.userCase = await req.locals.api.triggerEventWithData(
+        req.session.userCase.id,
+        formData,
+        eventName,
+        data
+      );
     } catch (err) {
       req.locals.logger.error('Error saving', err);
       req.session.errors = req.session.errors || [];
@@ -138,28 +178,89 @@ export class PostController<T extends AnyObject> {
     req: AppRequest<T>,
     res: Response,
     form: Form,
-    formData: Partial<Case>
+    formData: Partial<CaseWithId>
   ): Promise<void> {
-    // if (req?.session?.userCase) {
-    //   Object.assign(req?.session?.userCase, formData);
-    // } else {
-    //   const initData = { id: ' ', state: State.successAuthentication, serviceType: '', ...formData };
-    //   req.session.userCase = initData;
-    // }
+    const caseworkerUser = await getSystemUser();
+    const caseReference = formData.caseCode?.replace(/-/g, '');
+    const accessCode = formData.accessCode?.replace(/-/g, '');
+
     req.session.errors = form.getErrors(formData);
+
+    try {
+      if (!req.session.errors.length) {
+        const client = new CosApiClient(caseworkerUser.accessToken, 'http://localhost:3001');
+        const accessCodeValidated = await client.validateAccessCode(
+          caseReference as string,
+          accessCode as string,
+          caseworkerUser
+        );
+        if (accessCodeValidated === 'Linked') {
+          req.session.errors.push({ errorType: 'accesscodeAlreadyLinked', propertyName: 'accessCode' });
+        } else if (accessCodeValidated !== 'Valid') {
+          req.session.errors.push({ errorType: 'invalidAccessCode', propertyName: 'accessCode' });
+        }
+      }
+    } catch (err) {
+      console.log('Retrieving case failed with error: ' + err);
+      req.session.errors.push({ errorType: 'invalidReference', propertyName: 'caseCode' });
+    }
+
     if (req.session.errors.length) {
       req.session.accessCodeLoginIn = false;
     } else {
-      //make an api call to check if the caseId exists? and if it exists then set the case code
-      const initData = {
-        id: formData.caseCode || '',
-        state: State.successAuthentication,
-        serviceType: '',
-        ...formData,
-      };
-      req.session.userCase = initData;
       req.session.accessCodeLoginIn = true;
+      if (req?.session?.userCase) {
+        Object.assign(req?.session?.userCase, formData);
+      } else {
+        const initData = {
+          id: caseReference as string,
+          state: State.successAuthentication,
+          serviceType: '',
+          ...formData,
+        };
+        req.session.userCase = initData;
+      }
     }
+    this.redirect(req, res);
+  }
+
+  /**
+   * It takes a request, response, form and form data, and then assigns the form data to the user case
+   * in the session, and then sets the errors in the session to the errors from the form, and then
+   * filters the errors for save as draft, and then if there are errors in the session, it redirects to
+   * the same page, otherwise it redirects to the same page
+   * @param req - AppRequest<T> - this is the request object that is passed to the controller. It
+   * contains the session, the body, the query and the params.
+   * @param {Response} res - Response - the response object
+   * @param {Form} form - Form - the form object that is being used to render the page
+   * @param formData - The data that was submitted by the user
+   * @returns a promise.
+   */
+  private async onlyContinue(req: AppRequest<T>, res: Response, form: Form, formData: Partial<Case>): Promise<void> {
+    Object.assign(req.session.userCase, formData);
+    req.session.errors = form.getErrors(formData);
+    this.filterErrorsForSaveAsDraft(req);
+    if (req.session.errors.length) {
+      return this.redirect(req, res);
+    }
+
+    this.redirect(req, res);
+  }
+
+  private async getCaseList(req: AppRequest<T>, res: Response, form: Form, formData: Partial<Case>): Promise<void> {
+    req.session.errors = form.getErrors(formData);
+
+    this.filterErrorsForSaveAsDraft(req);
+
+    if (req.session.errors.length) {
+      return this.redirect(req, res);
+    }
+
+    const citizenUser = req.session.user;
+
+    const cosApiClient = new CosApiClient(citizenUser.accessToken, 'http://localhost:3001');
+    const caseDataFromCos = await cosApiClient.retrieveCasesByUserId(req.session.user);
+    console.log('retrieved casedata for case : ' + caseDataFromCos);
 
     this.redirect(req, res);
   }
