@@ -10,13 +10,13 @@ describe('AccessibilityStatementGetController', () => {
   const controller = new AccessibilityStatementGetController();
   const language = 'en';
 
-  test('Should render the accessibility statement page for adoption service', async () => {
+  test('Should render the accessibility statement page for private law service', async () => {
     const req = mockRequest();
     const res = mockResponse();
     await controller.get(req, res);
     const userCase = req.session.userCase;
 
-    expect(res.render).toBeCalledWith(expect.anything(), {
+    expect(res.render).toHaveBeenCalledWith(expect.anything(), {
       ...generatePageContent({
         language,
         pageContent: generateContent,
