@@ -1,7 +1,7 @@
 import mockUserCase from '../../../../test/unit/utils/mockUserCase';
-import { MIAM_START } from '../../urls';
+import { CONSENT, MIAM_START, PROCEEDINGS_COURT_PROCEEDINGS, PROCEEDINGS_START } from '../../urls';
 
-import { summaryList } from './utils';
+import { SummaryList, summaryList } from './utils';
 
 const enContent = {
   section: 'Check your details',
@@ -11,12 +11,32 @@ const enContent = {
   },
   keys: {
     miamStart: 'What is a Mediation Information and Assessment Meeting (MIAM)?',
+    applicationReceivedDate: 'When did you receive the application?',
+    invalidApplicationReceivedDate: 'When did you receive the application?',
+  },
+  fieldType: {
+    miamStart: 'string',
+    applicationReceivedDate: 'Date',
+    invalidApplicationReceivedDate: 'Date',
+    proceedingsStart: 'String',
+    proceedingsStartOrder: 'String',
+    emergencyOrderOptions: 'YesOrNo',
+    'emergencyOrder.caseNoDetails': 'String',
+    'emergencyOrder.orderDateDetails': 'Date',
+    'emergencyOrder.orderTimeDetails': 'String',
+    'emergencyOrder.currentOrderDetails': 'YesOrNo',
+    'emergencyOrder.issueOrderDetails': 'String',
   },
   errors: {},
 };
 
 const urls = {
   miamStart: MIAM_START,
+  proceedingsStart: PROCEEDINGS_START,
+  proceedingsStartOrder: PROCEEDINGS_START,
+  emergencyOrderOptions: PROCEEDINGS_COURT_PROCEEDINGS,
+  applicationReceivedDate: CONSENT,
+  invalidApplicationReceivedDate: CONSENT,
 };
 
 describe('common > summary > utils', () => {
@@ -25,26 +45,63 @@ describe('common > summary > utils', () => {
       {
         userCase: mockUserCase,
         expected: {
+          title: 'applicationDetails',
           rows: [
             {
               actions: {
                 items: [
                   {
-                    href: '/respondent/miam/miam-start',
+                    href: '/tasklistresponse/miam/miam-start',
                     text: 'Edit',
                     visuallyHiddenText: 'What is a Mediation Information and Assessment Meeting (MIAM)?',
                   },
                 ],
               },
               key: { text: 'What is a Mediation Information and Assessment Meeting (MIAM)?' },
-              value: { text: 'Yes' },
+              value: { html: 'Yes' },
+            },
+            {
+              actions: {
+                items: [
+                  {
+                    href: '/tasklistresponse/consent-to-application',
+                    text: 'Edit',
+                    visuallyHiddenText: 'When did you receive the application?',
+                  },
+                ],
+              },
+              key: { text: 'When did you receive the application?' },
+              value: { html: '11 March 2022' },
+            },
+            {
+              actions: {
+                items: [
+                  {
+                    href: '/tasklistresponse/consent-to-application',
+                    text: 'Edit',
+                    visuallyHiddenText: 'When did you receive the application?',
+                  },
+                ],
+              },
+              key: { text: 'When did you receive the application?' },
+              value: {},
             },
           ],
-          title: 'applicationDetails',
         },
       },
     ])('return correct summary list items when %#', ({ userCase, expected }) => {
-      expect(summaryList(enContent, userCase, urls, 'applicationDetails')).toStrictEqual(expected);
+      const result: SummaryList | undefined = summaryList(
+        enContent,
+        userCase,
+        urls,
+        'applicationDetails',
+        enContent.fieldType,
+        'en'
+      );
+      console.log(result);
+      expect(summaryList(enContent, userCase, urls, 'applicationDetails', enContent.fieldType, 'en')).toStrictEqual(
+        expected
+      );
     });
   });
 });
