@@ -1,8 +1,11 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { CaseWithId } from '../../../app/case/case';
-import { PageContent } from '../../../app/controller/GetController';
-//import * as Urls from '../../../../steps/urls';
+import dayjs from 'dayjs';
 
+import { CaseDate, CaseWithId } from '../../../app/case/case';
+import { PageContent } from '../../../app/controller/GetController';
+import { isDateInputInvalid } from '../../../app/form/validation';
+import { APPLICANT_TASK_LIST_URL, RESPONDENT_TASK_LIST_URL } from '../../../steps/urls';
 interface GovUkNunjucksSummary {
   key: {
     text?: string;
@@ -32,9 +35,10 @@ interface SummaryListRow {
   valueHtml?: string;
   changeUrl?: string;
   classes?: string;
+  caseLink?: string;
 }
 
-interface SummaryList {
+export interface SummaryList {
   title: string;
   rows: GovUkNunjucksSummary[];
 }
@@ -45,12 +49,12 @@ type SummaryListContent = PageContent & {
 };
 
 const getSectionSummaryList = (rows: SummaryListRow[], content: PageContent): GovUkNunjucksSummary[] => {
-  console.log(content);
+  console.log(content.title);
   return rows.map(item => {
     const changeUrl = item.changeUrl;
     return {
       key: { ...(item.key ? { text: item.key } : {}) },
-      value: { ...(item.value ? { text: item.value } : {}) },
+      value: { ...(item.value ? { html: item.value } : {}) },
       ...(changeUrl
         ? {
             actions: {
@@ -75,7 +79,9 @@ export const summaryList = (
   userCase: Partial<CaseWithId>,
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   urls: any,
-  sectionTitle?: string
+  sectionTitle?: string,
+  fieldTypes?: any,
+  language?: string
 ): SummaryList | undefined => {
   const summaryData: SummaryListRow[] = [];
   for (const key in keys) {
@@ -83,11 +89,12 @@ export const summaryList = (
     const url = urls[key];
     const row = {
       key: keyLabel,
-      value: userCase[key],
+      value: fieldTypes[key] === 'Date' ? getFormattedDate(userCase[key], language) : userCase[key],
       changeUrl: url,
     };
-
-    summaryData.push(row);
+    if (key !== 'applicant1SafeToCall') {
+      summaryData.push(row);
+    }
   }
 
   return {
@@ -95,3 +102,73 @@ export const summaryList = (
     rows: getSectionSummaryList(summaryData, content),
   };
 };
+
+export const summaryCaseList = (
+  userCaseList: Partial<CaseWithId>[],
+  sectionTitle?: string,
+  isRespondent?: boolean
+): SummaryList | undefined => {
+  const summaryData: SummaryListRow[] = [];
+  summaryData.push({ key: 'Case Name', value: '<h4>Case Status</h4>' });
+  for (const userCase of userCaseList) {
+    const id = userCase.id;
+    const name = userCase.applicantCaseName;
+    const state = userCase.state;
+    let caseUrl = '#';
+    if (userCase.caseTypeOfApplication === 'C100') {
+      if (!isRespondent) {
+        caseUrl = APPLICANT_TASK_LIST_URL + '/' + id;
+      } else {
+        caseUrl = RESPONDENT_TASK_LIST_URL + '/' + id;
+      }
+    } else if (userCase.caseTypeOfApplication === 'FL401') {
+      if (!isRespondent) {
+        caseUrl = APPLICANT_TASK_LIST_URL + '/' + id;
+      } else {
+        caseUrl = RESPONDENT_TASK_LIST_URL + '/' + id;
+      }
+    }
+    const row = {
+      key: name,
+      value: state,
+      changeUrl: id,
+      caseLink: caseUrl,
+    };
+
+    summaryData.push(row);
+  }
+
+  return {
+    title: sectionTitle || '',
+    rows: getSectionCaseList(summaryData),
+  };
+};
+
+const getSectionCaseList = (rows: SummaryListRow[]): GovUkNunjucksSummary[] => {
+  return rows.map(item => {
+    const changeUrl = item.changeUrl;
+    return {
+      key: { ...(item.key ? { text: item.key } : {}) },
+      value: { ...(item.value ? { html: item.value } : {}) },
+      ...(changeUrl
+        ? {
+            actions: {
+              items: [
+                {
+                  href: `${item.caseLink}`,
+                  text: `${item.changeUrl}`,
+                  visuallyHiddenText: `${item.changeUrl}`,
+                },
+              ],
+            },
+          }
+        : {}),
+      ...(item.classes ? { classes: item.classes } : {}),
+    };
+  });
+};
+
+export const getFormattedDate = (date: CaseDate | undefined, locale = 'en'): string =>
+  date && !isDateInputInvalid(date)
+    ? dayjs(`${date.day}-${date.month}-${date.year}`, 'D-M-YYYY').locale(locale).format('D MMMM YYYY')
+    : '';
