@@ -20,7 +20,7 @@ describe('ErrorController', () => {
 
     expect(logger.info.mock.calls[0][0]).toContain('404 Not Found: /request');
     expect(res.statusCode).toBe(404);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[404],
     });
@@ -35,7 +35,7 @@ describe('ErrorController', () => {
 
     expect(logger.error.mock.calls[0][0]).toContain('Bad request');
     expect(res.statusCode).toBe(err.status);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[400],
     });
@@ -50,7 +50,7 @@ describe('ErrorController', () => {
 
     expect(logger.error.mock.calls[0][0]).toContain('Bad request');
     expect(res.statusCode).toBe(500);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[500],
     });
@@ -65,7 +65,7 @@ describe('ErrorController', () => {
 
     expect(logger.error.mock.calls[0][0]).toContain('HTTPError: Bad request');
     expect(res.statusCode).toBe(400);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[400],
     });
@@ -79,7 +79,7 @@ describe('ErrorController', () => {
 
     expect(logger.error.mock.calls[0][0]).toContain('CSRF Token Failed');
     expect(res.statusCode).toBe(400);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[400],
     });
@@ -97,7 +97,7 @@ describe('ErrorController', () => {
     expect(logger.error.mock.calls[0][0]).toBe('Internal Server Error');
     expect(res.statusCode).toBe(500);
     expect(res.render).toHaveBeenCalledTimes(1);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en', userEmail: 'test@example.com' }),
       ...errorContent.en[500],
     });
@@ -128,7 +128,7 @@ describe('ErrorController', () => {
       error: 'invalid_grant',
     });
     expect(res.statusCode).toBe(500);
-    expect(res.render).toBeCalledWith('error/error', {
+    expect(res.render).toHaveBeenCalledWith('error/error', {
       ...generatePageContent({ language: 'en' }),
       ...errorContent.en[500],
     });
