@@ -1,5 +1,17 @@
 export const respondent_en = {
-  respondentYourDetails: 'Your details',
-  applicationDetails: 'Application detail',
-  respondentAdditionalInformation: 'Additional information',
+  yourDocuments: 'Your documents',
+  ordersFromTheCourt: 'Orders from the court',
+  yourResponse: 'Your response',
+  yourcourtHearings: 'Your court hearings',
+  theApplication: 'The application',
+  aboutYou: 'About you',
+};
+
+export const respondent_cy = {
+  yourDocuments: 'Your documents - welsh',
+  ordersFromTheCourt: 'Orders from the court - welsh',
+  yourResponse: 'Your response - welsh',
+  yourcourtHearings: 'Your court hearings - welsh',
+  theApplication: 'The application - welsh',
+  aboutYou: 'About you - welsh',
 };
