@@ -1,6 +1,28 @@
 export const respondent_tasklist_items_en = {
   keep_your_details_private: 'Keep your details private',
-  mediation_miam: 'Mediation(MIAM)',
-  international_factors: 'International element',
   confirm_or_edit_your_contact_details: 'Confirm or edit your contact details',
+  support_you_need_during_your_case: 'Support you need during your case',
+  check_the_application: 'Check the application (PDF)',
+  check_allegations_of_harm_and_violence: 'Check the allegations of harm and violence (PDF)',
+  respond_to_application: 'Respond to the application',
+  respond_to_allegations_of_harm_and_violence: 'Respond to the allegations of harm and violence',
+  check_details_of_your_court_hearings: 'Check details of your court hearings',
+  upload_document: 'Upload Documents',
+  view_all_orders_from_the_court: 'View all orders from the court',
+  view_all_documents: 'View all documents',
+};
+
+export const respondent_tasklist_items_cy = {
+  keep_your_details_private: 'Keep your details private - welsh',
+  confirm_or_edit_your_contact_details: 'Confirm or edit your contact details - welsh',
+  support_you_need_during_your_case: 'Support you need during your case - welsh',
+  check_the_application: 'Check the application (PDF) - welsh',
+  check_allegations_of_harm_and_violence: 'Check the allegations of harm and violence (PDF) - welsh',
+  respond_to_application: 'Respond to the application - welsh',
+  respond_to_allegations_of_harm_and_violence: 'Respond to the allegations of harm and violence - welsh',
+  check_details_of_your_court_hearings: 'Check details of your court hearings - welsh',
+  upload_document: 'Upload Documents - welsh',
+  view_all_orders_from_the_court: 'View all orders from the court - welsh',
+  view_all_documents: 'View all documents - welsh',
+  your_safety: 'Your safety - welsh',
 };
