@@ -16,8 +16,7 @@ exports.config = {
       },
       windowSize: '1280x960'
     },
-    PuppeteerHelpers: { require: './helpers/puppeterHelper.js' },
-    GeneralHelper: { require: './helpers/generalHelper.js' }
+    PuppeteerHelpers: { require: './helpers/puppeterHelper.js' }
   },
   plugins: {
     retryFailedStep: {
