@@ -14,7 +14,7 @@ module.exports = {
     caseTypeFullDesc: 'C100 & FL401 Applications'
   },
 
-  baseUrl: process.env.URL || 'https://prl-citizen-frontend-pr-80.service.core-compute-preview.internal/citizen-home',
+  baseUrl: process.env.PRL_CITIZEN_URL || 'https://prl-citizen-frontend-pr-213.service.core-compute-preview.internal/citizen-home',
   TestOutputDir: process.env.E2E_OUTPUT_DIR || './output',
   TestForAccessibility: process.env.TESTS_FOR_ACCESSIBILITY || false
 };
