@@ -5,6 +5,10 @@ const InternationalElement = require('./pages/InternationalElement');
 const generalHelper = require('./helpers/generalHelper');
 const Miam= require('./pages/Miam');
 
+const CurrentOrPreviousProceedings = require('./pages/CurrentOrPreviousProceedings');
+const ConsentToApplication = require('./pages/ConsentToApplication');
+const SupportYouNeed = require('./pages/SupportYouNeedDuringYourCase');
+
 module.exports = () => {
   return actor({
     
@@ -14,17 +18,23 @@ module.exports = () => {
     loginAsCitizen() {
       return Login.loginAsCitizen();
     },
-    homePage() {
-      return HomePage.clickRespondentLink();
-    },
     internationalElement() {
-      return InternationalElement.clickInternationalElement();
+      return InternationalElement.clickInternationalElementHappyPath();
+    },
+    currentOrPreviousProceedings() {
+      return CurrentOrPreviousProceedings.currentOrPreviousProceedingsHappyPath();
+    },
+    consentToApplication() {
+      return ConsentToApplication.consentToApplicationHappyPath();
     },
     amOnHistoryPageWithSuccessNotification() {
       return generalHelper.amOnHistoryPageWithSuccessNotification();
     },
     miam() {
       return Miam.addMIAM();
+
+    supportYouNeedDuringYourCase() {
+      return SupportYouNeed.supportYouNeedHappyPath();
     }
   });
 };
