@@ -1,8 +1,7 @@
 Feature('International Element');
 
-Scenario('International Element - basic journey @cross-browser', async I => {
+Scenario('International Element - basic journey @cross-browser', async ({ I }) => {
   await I.enterPinPageHappyPath();
   await I.loginAsCitizen();
-  await I.homePage();
   await I.internationalElement();
   }).retry({ retries: 3, minTimeout: 30000 });
