@@ -2,8 +2,8 @@ const path = require('path');
 
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 
-const packageJson = require.resolve('govuk-frontend/package.json');
-const root = path.resolve(packageJson, '..', 'govuk');
+const rootExport = require.resolve('govuk-frontend');
+const root = path.resolve(rootExport, '..');
 const sass = path.resolve(root, 'all.scss');
 const javascript = path.resolve(root, 'all.js');
 const components = path.resolve(root, 'components');
